Moved the accessors to an accessor directory-.
